+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.1 under development
 
 - Enh #265: Adapt summary data in debug collector (@rustamwin)
+- Bug #268: Remove debug console command from package configuration (@vjik)
 
 ## 4.0.0 February 25, 2025